add basic support for 'Go to definition'
