+++ cmd/lsp-server/find_node.go
@@ -2,72 +2,64 @@ package main
 
 import (
 	"github.com/escalier-lang/escalier/internal/ast"
-	"github.com/moznion/go-optional"
 )
 
 type Visitor struct {
 	Cursor ast.Location
-	Node   optional.Option[ast.Node]
+	Node   ast.Node
 }
 
 func (v *Visitor) VisitLit(l ast.Lit) bool {
 	if l.Span().Contains(v.Cursor) {
-		var node ast.Node = l
-		v.Node = optional.Some(node)
+		v.Node = l
 		return true
 	}
 	return false
 }
 func (v *Visitor) VisitPat(p ast.Pat) bool {
 	if p.Span().Contains(v.Cursor) {
-		var node ast.Node = p
-		v.Node = optional.Some(node)
+		v.Node = p
 		return true
 	}
 	return false
 }
 func (v *Visitor) VisitExpr(e ast.Expr) bool {
 	if e.Span().Contains(v.Cursor) {
-		var node ast.Node = e
-		v.Node = optional.Some(node)
+		v.Node = e
 		return true
 	}
 	return false
 }
 func (v *Visitor) VisitObjExprElem(e ast.ObjExprElem) bool {
 	if e.Span().Contains(v.Cursor) {
-		var node ast.Node = e
-		v.Node = optional.Some(node)
+		v.Node = e
 		return true
 	}
 	return false
 }
 func (v *Visitor) VisitStmt(s ast.Stmt) bool {
 	if s.Span().Contains(v.Cursor) {
-		var node ast.Node = s
-		v.Node = optional.Some(node)
+		v.Node = s
 		return true
 	}
 	return false
 }
 func (v *Visitor) VisitDecl(d ast.Decl) bool {
 	if d.Span().Contains(v.Cursor) {
-		var node ast.Node = d
-		v.Node = optional.Some(node)
+		v.Node = d
 		return true
 	}
 	return false
 }
 func (v *Visitor) VisitTypeAnn(t ast.TypeAnn) bool {
 	if t.Span().Contains(v.Cursor) {
-		var node ast.Node = t
-		v.Node = optional.Some(node)
+		v.Node = t
 		return true
 	}
 	return false
 }
 
-func findNodeInScript(script *ast.Script, loc ast.Location) optional.Option[ast.Node] {
+func findNodeInScript(script *ast.Script, loc ast.Location) ast.Node {
 	visitor := &Visitor{
 		Cursor: loc,
 		Node:   nil,
