+++ cmd/lsp-server/main.go
@@ -1,22 +1,14 @@
 package main
 
 import (
-	"context"
-	"encoding/json"
-	"errors"
 	"fmt"
 	"os"
-	"strings"
-	"time"
 
 	"github.com/tliron/glsp"
 	protocol "github.com/tliron/glsp/protocol_3_16"
 	glsp_server "github.com/tliron/glsp/server"
 
 	"github.com/escalier-lang/escalier/internal/ast"
-	"github.com/escalier-lang/escalier/internal/checker"
-	"github.com/escalier-lang/escalier/internal/compiler"
-	"github.com/escalier-lang/escalier/internal/parser"
 )
 
 const lsName = "escalier"
@@ -39,16 +31,16 @@ func main() {
 }
 
 type Server struct {
-	handler    protocol.Handler
-	documents  map[protocol.DocumentUri]protocol.TextDocumentItem
-	parsedDocs map[protocol.DocumentUri]*ast.Script
+	handler   protocol.Handler
+	documents map[protocol.DocumentUri]protocol.TextDocumentItem
+	astCache  map[protocol.DocumentUri]*ast.Script
 }
 
 func NewServer() *Server {
 	// nolint: exhaustruct
 	s := Server{
-		documents:  map[protocol.DocumentUri]protocol.TextDocumentItem{},
-		parsedDocs: map[protocol.DocumentUri]*ast.Script{},
+		documents: map[protocol.DocumentUri]protocol.TextDocumentItem{},
+		astCache:  map[protocol.DocumentUri]*ast.Script{},
 	}
 	// nolint: exhaustruct
 	s.handler = protocol.Handler{
@@ -58,12 +50,13 @@ func NewServer() *Server {
 		SetTrace:    s.setTrace,
 
 		// TextDocument
-		TextDocumentDidOpen:     s.textDocumentDidOpen,
-		TextDocumentDidChange:   s.textDocumentDidChange,
-		TextDocumentDeclaration: s.textDocumentDeclaration,
-		TextDocumentDefinition:  s.textDocumentDefinition,
-		TextDocumentHover:       s.textDocumentHover,
-		TextDocumentCodeAction:  s.textDocumentCodeAction,
+		TextDocumentDeclaration:    s.textDocumentDeclaration,
+		TextDocumentDefinition:     s.textDocumentDefinition,
+		TextDocumentTypeDefinition: s.textDocumentTypeDefinition,
+		TextDocumentDidOpen:        s.textDocumentDidOpen,
+		TextDocumentDidChange:      s.textDocumentDidChange,
+		TextDocumentHover:          s.textDocumentHover,
+		TextDocumentCodeAction:     s.textDocumentCodeAction,
 
 		// Workspace
 		WorkspaceExecuteCommand: s.workspaceExecuteCommand,
@@ -124,247 +117,4 @@ func (*Server) shutdown(context *glsp.Context) error {
 func (*Server) setTrace(context *glsp.Context, params *protocol.SetTraceParams) error {
 	protocol.SetTraceValue(params.Value)
 	return nil
-}
-
-// TODO: cache the results of the validation so that we can use them when responding
-// to hover and completion requests.
-func (server *Server) validate(lspContext *glsp.Context, uri protocol.DocumentUri, contents string) {
-	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
-	defer cancel()
-	p := parser.NewParser(ctx, parser.Source{
-		Path:     uri,
-		Contents: contents,
-	})
-	script, parseErrors := p.ParseScript()
-	server.parsedDocs[uri] = script
-
-	c := checker.NewChecker()
-	inferCtx := checker.Context{
-		Filename:   uri,
-		Scope:      checker.Prelude(),
-		IsAsync:    false,
-		IsPatMatch: false,
-	}
-	_, typeErrors := c.InferScript(inferCtx, script)
-
-	diagnotics := []protocol.Diagnostic{}
-	for _, err := range parseErrors {
-		severity := protocol.DiagnosticSeverityError
-		source := "escalier"
-		diagnotics = append(diagnotics, protocol.Diagnostic{
-			Range: protocol.Range{
-				Start: protocol.Position{
-					Line:      protocol.UInteger(err.Span.Start.Line - 1),
-					Character: protocol.UInteger(err.Span.Start.Column - 1),
-				},
-				End: protocol.Position{
-					Line:      protocol.UInteger(err.Span.End.Line - 1),
-					Character: protocol.UInteger(err.Span.End.Column - 1),
-				},
-			},
-			Severity:           &severity,
-			Code:               nil,
-			CodeDescription:    nil,
-			Source:             &source,
-			Message:            err.Message,
-			Tags:               nil,
-			RelatedInformation: nil,
-			Data:               nil,
-		})
-	}
-
-	for _, err := range typeErrors {
-		severity := protocol.DiagnosticSeverityError
-		source := "escalier"
-		diagnotics = append(diagnotics, protocol.Diagnostic{
-			Range: protocol.Range{
-				Start: protocol.Position{
-					Line:      0,
-					Character: 0,
-				},
-				End: protocol.Position{
-					Line:      0,
-					Character: 0,
-				},
-			},
-			Severity:           &severity,
-			Code:               nil,
-			CodeDescription:    nil,
-			Source:             &source,
-			Message:            err.Message,
-			Tags:               nil,
-			RelatedInformation: nil,
-			Data:               nil,
-		})
-	}
-
-	go lspContext.Notify(protocol.ServerTextDocumentPublishDiagnostics, &protocol.PublishDiagnosticsParams{
-		URI:         uri,
-		Diagnostics: diagnotics,
-		Version:     nil,
-	})
-}
-
-func (s *Server) textDocumentDidOpen(context *glsp.Context, params *protocol.DidOpenTextDocumentParams) error {
-	s.documents[params.TextDocument.URI] = params.TextDocument
-	if params.TextDocument.LanguageID == "escalier" {
-		s.validate(context, params.TextDocument.URI, params.TextDocument.Text)
-	}
-	return nil
-}
-
-func (s *Server) textDocumentDidChange(context *glsp.Context, params *protocol.DidChangeTextDocumentParams) error {
-	doc := s.documents[params.TextDocument.URI]
-
-	for _, change := range params.ContentChanges {
-		switch change := change.(type) {
-		case protocol.TextDocumentContentChangeEvent:
-			return fmt.Errorf("incremental changes not supported")
-		case protocol.TextDocumentContentChangeEventWhole:
-			s.documents[params.TextDocument.URI] = protocol.TextDocumentItem{
-				URI:        params.TextDocument.URI,
-				LanguageID: doc.LanguageID,
-				Version:    params.TextDocument.Version,
-				Text:       change.Text,
-			}
-		}
-	}
-
-	if doc.LanguageID == "escalier" {
-		for _, _change := range params.ContentChanges {
-			change := _change.(protocol.TextDocumentContentChangeEventWhole)
-			s.validate(context, params.TextDocument.URI, change.Text)
-		}
-	}
-	return nil
-}
-
-func (*Server) textDocumentDeclaration(context *glsp.Context, params *protocol.DeclarationParams) (any, error) {
-	fmt.Fprintf(os.Stderr, "textDocumentDeclaration - uri = %s\n", params.TextDocument.URI)
-	err := fmt.Errorf("textDocument/declaration not implemented yet")
-	return nil, err
-}
-
-func (*Server) textDocumentDefinition(context *glsp.Context, params *protocol.DefinitionParams) (any, error) {
-	fmt.Fprintf(os.Stderr, "textDocumentDefinition - uri = %s\n", params.TextDocument.URI)
-	err := fmt.Errorf("textDocument/definition not implemented yet")
-	return nil, err
-}
-
-func (server *Server) textDocumentHover(context *glsp.Context, params *protocol.HoverParams) (*protocol.Hover, error) {
-	fmt.Fprintf(os.Stderr, "textDocumentHover - uri = %s\n", params.TextDocument.URI)
-	fmt.Fprintf(os.Stderr, "textDocumentHover - position = line:%d, column:%d\n", params.Position.Line, params.Position.Character)
-
-	value := fmt.Sprintf(
-		"textDocumentHover - position = line:%d, column:%d\n",
-		params.Position.Line,
-		params.Position.Character,
-	)
-
-	script := server.parsedDocs[params.TextDocument.URI]
-	if script != nil {
-		loc := ast.Location{
-			Line:   int(params.Position.Line),
-			Column: int(params.Position.Character),
-		}
-		node := findNodeInScript(script, loc)
-		node.IfSome(func(node ast.Node) {
-			switch node := node.(type) {
-			case ast.Expr:
-				if node.InferredType() != nil {
-					value = "`" + node.InferredType().String() + "`"
-				}
-			case ast.Pat:
-				if node.InferredType() != nil {
-					value = "`" + node.InferredType().String() + "`"
-				}
-			default:
-				// do nothing
-			}
-		})
-	}
-
-	hover := protocol.Hover{
-		Contents: protocol.MarkupContent{
-			Kind:  protocol.MarkupKindMarkdown,
-			Value: value,
-		},
-		Range: nil,
-	}
-	return &hover, nil
-}
-
-func addr[T any](x T) *T {
-	return &x
-}
-
-func (*Server) textDocumentCodeAction(context *glsp.Context, params *protocol.CodeActionParams) (any, error) {
-	compileAction := protocol.CodeAction{
-		Title:       "Compile",
-		Kind:        addr("compile"),
-		Diagnostics: []protocol.Diagnostic{},
-		IsPreferred: nil,
-		Disabled:    nil,
-		Edit:        nil, // Require the client to make a workspace/executeCommand request
-		Command: &protocol.Command{
-			Title:     "Compile",
-			Command:   "compile",
-			Arguments: []any{},
-		},
-		Data: nil,
-	}
-
-	codeActions := []protocol.CodeAction{compileAction}
-
-	return codeActions, nil
-}
-
-func (s *Server) workspaceExecuteCommand(context *glsp.Context, params *protocol.ExecuteCommandParams) (any, error) {
-	if params.Command != "compile" {
-		return nil, fmt.Errorf("unknown command: %s", params.Command)
-	}
-
-	if len(params.Arguments) != 1 {
-		return nil, fmt.Errorf("invalid arguments: %v", params.Arguments)
-	}
-
-	uri, ok := params.Arguments[0].(protocol.DocumentUri)
-	if !ok {
-		return nil, fmt.Errorf("invalid argument: %v", params.Arguments[0])
-	}
-
-	doc, ok := s.documents[uri]
-	if !ok {
-		fmt.Fprintf(os.Stderr, "document not found: %s\n", uri)
-		return nil, fmt.Errorf("document not found: %s", uri)
-	}
-
-	if doc.LanguageID != "escalier" {
-		return nil, fmt.Errorf("unsupported language: %s", doc.LanguageID)
-	}
-
-	source := parser.Source{
-		Path:     uri,
-		Contents: doc.Text,
-	}
-
-	output := compiler.Compile(source)
-
-	if len(output.ParseErrors) > 0 {
-		errorsJSON, err := json.Marshal(output.ParseErrors)
-		if err != nil {
-			return nil, fmt.Errorf("failed to marshal compilation errors: %v", err)
-		}
-		return nil, errors.New(string(errorsJSON))
-	}
-
-	// TODO: include errors in the response
-	response := protocol.TextDocumentItem{
-		URI:        strings.TrimSuffix(uri, ".esc") + ".js",
-		LanguageID: "javascript",
-		Version:    0,
-		Text:       output.JS,
-	}
-
-	return response, nil
 }
