+++ internal/ast/binding.go
@@ -2,7 +2,8 @@ package ast
 
 type BindingSource interface {
 	isBindingSource()
+	Node
 }
 
-func (*Ident) isBindingSource()    {}
+func (*Ident) isBindingSource()    {} // Used for FuncDecls and TypeDecls
 func (*IdentPat) isBindingSource() {}
